I've renamed cohort and students page names and added Admin to it and created a new folder for my Cohorts Page
